+++ forge19/src/main/java/com/vecoo/extralib/storage/IServerJsonProvider.java
@@ -1,0 +1,11 @@
+package com.vecoo.extralib.storage;
+
+public interface IServerJsonProvider<T> {
+    T getStorage();
+
+    void updateStorage(T storage);
+
+    void write();
+
+    void init();
+}