+++ forge20/src/main/java/com/vecoo/extralib/storage/IConfigProvider.java
@@ -1,0 +1,7 @@
+package com.vecoo.extralib.storage;
+
+public interface IConfigProvider {
+    boolean write();
+
+    void init();
+}