+++ neoforge21/src/main/java/com/vecoo/extralib/player/UtilPlayer.java
@@ -11,12 +11,14 @@ import net.minecraft.world.inventory.InventoryMenu;
 import net.minecraft.world.item.ItemStack;
 import net.neoforged.neoforge.common.UsernameCache;
 
+import javax.annotation.Nullable;
 import java.util.Map;
 import java.util.Objects;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
 public class UtilPlayer {
+    @Nullable
     public static UUID getUUID(String playerName) {
         return UsernameCache.getMap().entrySet().stream().collect(Collectors.toMap(Map.Entry::getValue, Map.Entry::getKey)).get(playerName);
     }